pyproject.toml
[tool.poetry]
name = "announce"
version = "0.1.0"
description = "Code used to announce pyjaipur events"
authors = ["arjoonn sharma <arjoonn.94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.23.0"
requests-oauthlib = "^1.3.0"
google-api-python-client = "^1.8.4"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
pendulum = "^2.1.0"
oauth2client = "^4.1.3"
Jinja2 = "^2.11.2"
bottle = "^0.12.18"
bottle-tools = "^2019.12.22"
sqlalchemy = "^1.3.17"
psycopg2-binary = "^2.8.5"
python-telegram-bot = "^12.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
